requirements.txt
Pillow>=10.0
PyYAML>=6.0
google-api-python-client>=2.100
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
Flask>=3.0
